% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdee.R
\name{tdee}
\alias{tdee}
\title{A TDEE calculator for R.}
\usage{
tdee(height, weight, age, sex, activity, aim)
}
\arguments{
\item{height}{Your height in centimetres. Returns a warning message to check unit of measurement if over 250 or under 100.}

\item{weight}{Your weight in kilograms. Returns a warning message to check units of measurement if over 250 or under 30.}

\item{age}{Your age in years. Returns a warning message to check units of measurement if over 100.}

\item{sex}{Your biological sex. Throws an error if not 'male' or 'female'.}

\item{activity}{Your activity level. Are you 'sedentary', or 'lightly', 'moderately', 'very' or 'extremely' active? Throws an error if not one of these options (do not include the word 'active', just the adverbs).}

\item{aim}{Your weight aim. Do you want to 'lose', 'maintain', or 'gain' weight? Throws an error if not one of these options}
}
\value{
A dataframe with columns bmr, tdee, and calories
}
\description{
Calculate your basal metabolic rate (BMR), total daily energy expenditure (TDEE), and calorie goals to lose, maintain, or gain weight.
}
\details{
\code{tdee} calculates your BMR, TDEE, and calorie requirements based on your height, weight, age, sex, activity levels and weight aims. The calculations are based on this post: https://steelfitusa.com/2018/10/calculate-tdee/
}
\examples{
# my own calculation
tdee(height = 182,
     weight = 84,
     age = 23,
     sex = "male",
     activity = "moderately",
     aim = "gain")
# alternatively without arg names
tdee(182, 84, 23, "male", "moderately", "gain")
# a 150 cm, thirty-year-old woman weighing ninety kilograms, who is
# sedentary and wants to lose weight
tdee(height = 150,
     weight = 90,
     age = 30,
     sex = "female",
     activity = "sedentary",
     aim = "lose")
}
